[feat] 기본 user, dish, comment 구조개발

+++ backend/src/services/dish.service.js
@@ -1,22 +1,20 @@
 import prisma from '../utils/prisma.js';
-
-// 공통 select (중복 제거)
-const dishSelect = {
-  id: true,
-  title: true,
-  description: true,
-  imageUrl: true,
-  createdAt: true,
-  updatedAt: true,
-};
+import { getOneByIdOrFail, userSelect } from '../utils/index.js';
 
 // 모든 게시글 조회
 export const getDishes = async (page = 1, limit = 10) => {
   const skip = (page - 1) * limit;
+
   const dishes = await prisma.dish.findMany({
     skip,
     take: limit,
-    include: { user: true, comments: true, likes: true },
+    include: {
+      user: {
+        select: userSelect,
+      },
+      comments: true,
+      likes: true,
+    },
     orderBy: { createdAt: 'desc' },
   });
   return dishes;
@@ -26,14 +24,20 @@ export const getDishes = async (page = 1, limit = 10) => {
 export const getDishById = async (id) => {
   const dish = await prisma.dish.findUnique({
     where: { id },
-    include: { user: true, comments: true, likes: true },
+    include: {
+      user: {
+        select: userSelect,
+      },
+      comments: true,
+      likes: true,
+    },
   });
   if (!dish) throw new Error('게시글을 찾을 수 없습니다.');
   return dish;
 };
 
 // 게시글 생성
-export const createDish = async (data) => {
+export const createDish = async (userId, data) => {
   const { title, description, imageUrl } = data;
 
   const dish = await prisma.dish.create({
@@ -41,15 +45,25 @@ export const createDish = async (data) => {
       title,
       description,
       imageUrl,
+      userId,
     },
-    include: { user: true, comments: true, likes: true },
+    include: {
+      user: {
+        select: userSelect,
+      },
+    },
   });
   return dish;
 };
 
 // 게시글 수정
-export const updateDish = async (id, data) => {
+export const updateDish = async (id, userId, data) => {
   const { title, description, imageUrl } = data;
+  // 게시물이 존재하는지 확인
+  const dishData = await getOneByIdOrFail(prisma.dish, id, '게시글');
+  if (dishData.userId !== userId) {
+    throw new Error('게시글을 수정할 권한이 없습니다.');
+  }
 
   const updateData = {
     ...(title && { title }),
@@ -60,13 +74,25 @@ export const updateDish = async (id, data) => {
   const dish = await prisma.dish.update({
     where: { id },
     data: updateData,
-    include: { user: true, comments: true, likes: true },
+    include: {
+      user: {
+        select: userSelect,
+      },
+      comments: true,
+      likes: true,
+    },
   });
   return dish;
 };
 
 // 게시글 삭제
-export const deleteDish = async (id) => {
+export const deleteDish = async (id, userId) => {
+  // 게시물이 존재하는지 확인
+  const dishData = await getOneByIdOrFail(prisma.dish, id, '게시글');
+  if (dishData.userId !== userId) {
+    throw new Error('게시글을 삭제할 권한이 없습니다.');
+  }
+
   await prisma.dish.delete({
     where: { id },
   });
